Struct ddc_macos::Monitor

pub struct Monitor { /* private fields */ }

A handle to an attached monitor that allows the use of DDC/CI operations.

Implementations

impl Monitor

pub fn enumerate() -> Result<Vec<Self>, Error>

Enumerate all connected physical monitors returning Vec

pub fn description(&self) -> String

Physical monitor description string. If it cannot get the product’s name it will use the vendor number and model number to form a description

pub fn serial_number(&self) -> Option<String>

Serial number for this Monitor

pub fn product_name(&self) -> Option<String>

Product name for this Monitor, if available

pub fn edid(&self) -> Option<Vec<u8>>

Returns Extended display identification data (EDID) for this Monitor as raw bytes data

pub fn handle(&self) -> CGDisplay

CoreGraphics display handle for this monitor